Cp2: Moving into GDScript for faster prototyping
I switched from image to Raylib for realtime prototyping. But I wondered, why not use Python instead? So I don't had to fighting with static typing. But why use Python when GDScript run faster and it already include the both 2D and 3D graphics engine. So that's why I use GDScript to prototype it.